+++ src/commands/workspace/mod.rs
@@ -1,7 +1,7 @@
+mod config;
 mod info;
 mod run;
-mod set;
 
+pub use config::config;
 pub use info::info;
 pub use run::run;
-pub use set::{set, unset};